refactor: id 값 null 체크 코드 추가

+++ src/main/java/com/example/board/comment/controller/CommentController.java
@@ -35,6 +35,11 @@ public class CommentController {
     // 댓글 삭제
     @DeleteMapping("/{commentId}")
     public ResponseEntity<String> deleteComment(@PathVariable(name = "commentId") Long commentId) {
+
+        if (commentId == null) {
+            return ResponseEntity.badRequest().body("commentId가 유효하지 않습니다.");
+        }
+
         commentService.deleteComment(commentId);
         return ResponseEntity.ok("댓글이 삭제되었습니다.");
     }
