+++ src/main/java/com/example/board/post/controller/PostController.java
@@ -1,10 +1,8 @@
 package com.example.board.post.controller;
 
-import com.example.board.board.entity.Board;
 import com.example.board.post.dto.request.PostEditRequest;
 import com.example.board.post.dto.request.PostOneRequest;
 import com.example.board.post.dto.request.PostRequest;
-import com.example.board.post.dto.response.PostResponse;
 import com.example.board.post.entity.Post;
 import com.example.board.post.service.PostService;
 import java.net.URLDecoder;
@@ -47,6 +45,10 @@ public class PostController {
     // 게시글 삭제
     @DeleteMapping("/{postId}")
     public ResponseEntity<String> deletePost(@PathVariable(name = "postId") Long postId) {
+        if (postId == null) {
+            return ResponseEntity.badRequest().body("postId가 유효하지 않습니다.");
+        }
+
         postService.deletePost(postId);
         return ResponseEntity.ok("게시글이 삭제되었습니다.");
     }
@@ -55,6 +57,7 @@ public class PostController {
     @PutMapping("/{postid}")
     public ResponseEntity<Post> updatePost(@PathVariable(name = "postid") Long postId,
                                            @RequestBody PostEditRequest postEditRequest) {
+
         Post updatedPost = postService.updatePost(postId, postEditRequest);
         return ResponseEntity.ok(updatedPost);
     }
